refactor: split bridge.ts into domain-specific route modules
Break the 880-line monolith into src/routes/bridge/ with 9 domain
files (ledger, credentials, finance, plaid, mercury, books, assets,
scrape, status) and a barrel index. No behavior changes.
